Jhamela

+++ MyApp/Areas/Admin/Controllers/ProductController.cs
@@ -111,7 +111,7 @@ namespace MyApp.Web.Areas.Admin.Controllers
                     {
                         file.CopyTo(fileStream);
                     }
-                    productVM.Product.ImageUrl = @"\Image\" + fileName;
+                    productVM.Product.ImageUrl = @"/Images/" + fileName;
                 }
                 if (productVM.Product.Id == 0)
                 {
@@ -148,16 +148,21 @@ namespace MyApp.Web.Areas.Admin.Controllers
             }
             else
             {
-                var oldImagePath = Path.Combine(_webHostEnvironment.WebRootPath, product.ImageUrl.TrimStart('\\'));
-                if (System.IO.File.Exists(oldImagePath))
+                if (product.ImageUrl != null)
                 {
-                    System.IO.File.Delete(oldImagePath);
+                    var oldImagePath = Path.Combine(_webHostEnvironment.WebRootPath, product.ImageUrl.TrimStart('\\'));
+                    if (System.IO.File.Exists(oldImagePath))
+                    {
+                        System.IO.File.Delete(oldImagePath);
+                    }
                 }
+               
 
                 _appDbContext.Product.Delete(product);
                 TempData["error"] = "Deleted";
                 _appDbContext.Save();
-                return Json(new { success = true, message = "Product Deleted" });
+                return Json(new { success = true, message = "Product Deleted !!!" });
+                //return RedirectToAction("Index");
             }
         }
         #endregion
